changed everything to float, added sprite, added levelpiece, etc

+++ Position.java
@@ -1,6 +1,6 @@
 public class Position {
-  public final int x,y;
-  public Position(int x,int y) { this.x=x; this.y=y; }
+  public final float x,y;
+  public Position(float x,float y) { this.x=x; this.y=y; }
   public Position() { this(0,0); }
   public Position(Position p) { this(p.x,p.y); }
 
@@ -8,7 +8,7 @@ public class Position {
   Position sub(Position p) { return new Position(x-p.x,y-p.y); }
   Position mul(Position p) { return new Position(x*p.x,y*p.y); }
   Position div(Position p) { return new Position(x/p.x,y/p.y); }
-  Position mul(int m)      { return mul( new Position(m,m) ); }
-  Position div(int m)      { return div( new Position(m,m) ); }
+  Position mul(float m)      { return mul( new Position(m,m) ); }
+  Position div(float m)      { return div( new Position(m,m) ); }
   boolean  equ(Position p) { return x==p.x&&y==p.y; }
 }
